Added Competition API

+++ src/main/java/nz/roag/archerylogbook/rest/ArcherService.java
@@ -8,11 +8,13 @@ import nz.roag.archerylogbook.db.model.Club;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 public class ArcherService {
@@ -25,15 +27,17 @@ public class ArcherService {
     @Autowired
     private ClubRepository clubRepository;
 
-    public List<Archer> listAllArchers() {
+    public Page<Archer> listAllArchers(int page, int size) {
         logger.debug("Getting list of all archers");
-        return archerRepository.findAll(Sort.by("lastName").ascending()
-                .and(Sort.by("firstName").ascending()));
+        return archerRepository.findByArchived(false,
+                PageRequest.of(page, size,
+                        Sort.by("lastName").ascending().and(Sort.by("firstName").ascending())));
     }
 
-    public List<Archer> listArchersByClub(long clubId) {
+    public Page<Archer> listArchersByClub(long clubId, int page, int size) {
         logger.debug("Getting list of archers for clubId {}", clubId);
-        return archerRepository.findByClubIdOrderByLastNameAsc(clubId);
+        return archerRepository.findByClubIdOrderByLastNameAsc(clubId, PageRequest.of(page, size,
+                Sort.by("lastName").ascending().and(Sort.by("firstName").ascending())));
     }
 
     @Transactional
@@ -61,7 +65,22 @@ public class ArcherService {
     @Transactional
     public void deleteArcher(long id) {
         logger.warn("Deleting archer with id {}", id);
-        archerRepository.deleteById(id);
+        Optional<Archer> archerOpt = archerRepository.findById(id);
+        if (archerOpt.isPresent()) {
+            if (archerOpt.get().getRoundList().isEmpty() && archerOpt.get().getBowList().isEmpty()
+                    && archerOpt.get().getCompetitionList().isEmpty()) {
+                //no child records in DB, so we can directly delete the archer
+                archerRepository.deleteById(id);
+            } else {
+                //archer has a history, so we archive it instead of deletion
+                var archer = archerOpt.get();
+                archer.setArchived(true);
+                archer.getBowList().forEach(bow -> bow.setArchived(true));
+                archer.getRoundList().forEach(round -> round.setArchived(true));
+                archer.getCompetitionList().forEach(cmpt -> cmpt.setArchived(true));
+                archerRepository.save(archer);
+            }
+        }
     }
 
     @Transactional
